Non DL model added
